# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Velmora Luxury Perfume E-Commerce API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  message: string;
}

export interface SuccessResponse {
  success: boolean;
  message: string;
}

export interface InscriptionBody {
  nom: string;
  email: string;
  /** @minLength 6 */
  motDePasse: string;
}

export interface ConnexionBody {
  email: string;
  motDePasse: string;
}

export type UtilisateurRole =
  (typeof UtilisateurRole)[keyof typeof UtilisateurRole];

export const UtilisateurRole = {
  client: "client",
  admin: "admin",
} as const;

export interface Utilisateur {
  id: number;
  nom: string;
  email: string;
  role: UtilisateurRole;
  createdAt: string;
}

export interface AuthResponse {
  token: string;
  utilisateur: Utilisateur;
}

export type ProduitCategorie =
  (typeof ProduitCategorie)[keyof typeof ProduitCategorie];

export const ProduitCategorie = {
  homme: "homme",
  femme: "femme",
  unisexe: "unisexe",
} as const;

export interface Produit {
  id: number;
  nom: string;
  description: string;
  prix: number;
  imageUrl: string;
  categorie: ProduitCategorie;
  enVedette: boolean;
  nombreVentes: number;
  createdAt: string;
}

export type ProduitBodyCategorie =
  (typeof ProduitBodyCategorie)[keyof typeof ProduitBodyCategorie];

export const ProduitBodyCategorie = {
  homme: "homme",
  femme: "femme",
  unisexe: "unisexe",
} as const;

export interface ProduitBody {
  nom: string;
  description: string;
  prix: number;
  imageUrl: string;
  categorie: ProduitBodyCategorie;
  enVedette?: boolean;
}

export interface ProduitsPage {
  produits: Produit[];
  total: number;
  page: number;
  totalPages: number;
}

export type CatalogueStatsParCategorie = {
  homme: number;
  femme: number;
  unisexe: number;
};

export interface CatalogueStats {
  total: number;
  parCategorie: CatalogueStatsParCategorie;
}

export interface PanierItem {
  produitId: number;
  quantite: number;
  produit: Produit;
}

export interface Panier {
  items: PanierItem[];
  total: number;
  nombreArticles: number;
}

export interface PanierItemBody {
  produitId: number;
  quantite?: number;
}

export interface ModifierQuantiteBody {
  /** @minimum 0 */
  quantite: number;
}

export interface CommandeItem {
  id: number;
  produitId: number;
  quantite: number;
  prixUnitaire: number;
  produit: Produit;
}

export type CommandeStatut =
  (typeof CommandeStatut)[keyof typeof CommandeStatut];

export const CommandeStatut = {
  en_attente: "en_attente",
  confirmee: "confirmee",
  expediee: "expediee",
  livree: "livree",
  annulee: "annulee",
} as const;

export interface Commande {
  id: number;
  userId: number;
  total: number;
  statut: CommandeStatut;
  items: CommandeItem[];
  utilisateur?: Utilisateur;
  createdAt: string;
}

export type UpdateStatutBodyStatut =
  (typeof UpdateStatutBodyStatut)[keyof typeof UpdateStatutBodyStatut];

export const UpdateStatutBodyStatut = {
  en_attente: "en_attente",
  confirmee: "confirmee",
  expediee: "expediee",
  livree: "livree",
  annulee: "annulee",
} as const;

export interface UpdateStatutBody {
  statut: UpdateStatutBodyStatut;
}

export type AdminStatsCommandesParStatut = {
  en_attente: number;
  confirmee: number;
  expediee: number;
  livree: number;
  annulee: number;
};

export interface AdminStats {
  totalCommandes: number;
  chiffreAffaires: number;
  totalClients: number;
  totalProduits: number;
  commandesParStatut: AdminStatsCommandesParStatut;
  commandesRecentes: Commande[];
}

export type GetProduitsParams = {
  recherche?: string;
  categorie?: GetProduitsCategorie;
  prixMin?: number;
  prixMax?: number;
  page?: number;
  limite?: number;
};

export type GetProduitsCategorie =
  (typeof GetProduitsCategorie)[keyof typeof GetProduitsCategorie];

export const GetProduitsCategorie = {
  homme: "homme",
  femme: "femme",
  unisexe: "unisexe",
} as const;