# lib/api-zod/src/generated/types/commandeItem.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Velmora Luxury Perfume E-Commerce API
 * OpenAPI spec version: 0.1.0
 */
import type { Produit } from "./produit";

export interface CommandeItem {
  id: number;
  produitId: number;
  quantite: number;
  prixUnitaire: number;
  produit: Produit;
}