# lib/api-zod/src/generated/types/modifierQuantiteBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Velmora Luxury Perfume E-Commerce API
 * OpenAPI spec version: 0.1.0
 */

export interface ModifierQuantiteBody {
  /** @minimum 0 */
  quantite: number;
}