# lib/api-zod/src/generated/types/panier.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Velmora Luxury Perfume E-Commerce API
 * OpenAPI spec version: 0.1.0
 */
import type { PanierItem } from "./panierItem";

export interface Panier {
  items: PanierItem[];
  total: number;
  nombreArticles: number;
}