# lib/api-zod/src/generated/types/panierItem.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Velmora Luxury Perfume E-Commerce API
 * OpenAPI spec version: 0.1.0
 */
import type { Produit } from "./produit";

export interface PanierItem {
  produitId: number;
  quantite: number;
  produit: Produit;
}